apps/api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
supabase>=2.3.0
httpx>=0.26.0
python-multipart>=0.0.9
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
celery>=5.3.6
redis>=5.0.1
pgvector>=0.2.4
beautifulsoup4>=4.12.3
youtube-transcript-api>=0.6.2
yt-dlp>=2024.0.0
